# backend/docs/swagger-es.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LoL API v1",
    "version": "1.0.0",
    "description": "API para campeones de League of Legends y funcionalidades de usuario.\n\nLas rutas protegidas requieren un Session Token en el header Authorization:\nAuthorization: Bearer <sessionToken>\n\nNota: este token no tiene por qué ser un JWT. Es un token de sesión generado por la API."
  },
  "servers": [
    {
      "url": "http://localhost:3010",
      "description": "Servidor local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "SessionToken",
        "description": "Usar el sessionToken devuelto por /api/v1/auth/login. Ejemplo: Authorization: Bearer <sessionToken>"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "error" },
          "message": { "type": "string", "example": "Mensaje de error" },
          "details": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "Skin": {
        "type": "object",
        "required": ["name", "splashUrl", "loadingUrl"],
        "properties": {
          "name": { "type": "string", "example": "Fiddlesticks Star Nemesis" },
          "splashUrl": {
            "type": "string",
            "format": "uri",
            "description": "Imagen horizontal grande (desktop)",
            "example": "https://ddragon.leagueoflegends.com/cdn/img/champion/splash/Fiddlesticks_10.jpg"
          },
          "loadingUrl": {
            "type": "string",
            "format": "uri",
            "description": "Imagen vertical (mobile / loading)",
            "example": "https://ddragon.leagueoflegends.com/cdn/img/champion/loading/Fiddlesticks_10.jpg"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "description": "Legacy (compatibilidad). Normalmente igual a splashUrl.",
            "example": "https://ddragon.leagueoflegends.com/cdn/img/champion/splash/Fiddlesticks_10.jpg"
          }
        }
      },
      "ChampionSummary": {
        "type": "object",
        "required": ["id", "name", "positions", "region"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "title": { "type": "string" },
          "iconUrl": { "type": "string", "format": "uri" },
          "splashUrl": { "type": "string", "format": "uri" },
          "region": { "type": "string" },
          "positions": { "type": "array", "items": { "type": "string" } }
        }
      },
      "Champion": {
        "allOf": [
          { "$ref": "#/components/schemas/ChampionSummary" },
          {
            "type": "object",
            "properties": {
              "seed": { "type": "boolean" },
              "owner": { "type": "string" },
              "roles": { "type": "array", "items": { "type": "string" } },
              "skins": { "type": "array", "items": { "$ref": "#/components/schemas/Skin" } },
              "lore": { "type": "string" },
              "allytips": { "type": "array", "items": { "type": "string" } },
              "enemytips": { "type": "array", "items": { "type": "string" } },
              "info": { "type": "object" },
              "stats": { "type": "object" },
              "abilities": { "type": "object" }
            }
          }
        ]
      },
      "PaginatedChampionSummaries": {
        "type": "object",
        "properties": {
          "meta": {
            "type": "object",
            "properties": {
              "total": { "type": "integer" },
              "page": { "type": "integer" },
              "limit": { "type": "integer" },
              "totalPages": { "type": "integer" }
            }
          },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ChampionSummary" }
          }
        }
      },
      "CreateChampionBody": {
        "type": "object",
        "required": ["id", "name", "region", "positions"],
        "properties": {
          "id": { "type": "string", "example": "MyUniqueChamp" },
          "name": { "type": "string", "example": "Mi Campeón" },
          "title": { "type": "string", "example": "El Legendario" },
          "region": { "type": "string", "example": "Runaterra" },
          "positions": {
            "type": "array",
            "items": { "type": "string", "enum": ["Top", "Jungle", "Mid", "ADC", "Support"] },
            "example": ["Mid"]
          },
          "roles": { "type": "array", "items": { "type": "string" }, "example": ["Mage"] },
          "lore": { "type": "string", "example": "Un campeón custom." },
          "allytips": { "type": "array", "items": { "type": "string" } },
          "enemytips": { "type": "array", "items": { "type": "string" } }
        }
      },
      "UpdateChampionBody": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "title": { "type": "string" },
          "region": { "type": "string" },
          "positions": { "type": "array", "items": { "type": "string" } },
          "roles": { "type": "array", "items": { "type": "string" } },
          "lore": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/champions": {
      "get": {
        "tags": ["Campeones"],
        "summary": "Listar campeones seed (paginado)",
        "description": "Devuelve una lista paginada de campeones oficiales (seed). Usar 'page' y 'limit'.",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "en", "enum": ["en", "es"] },
            "description": "Idioma de textos del campeón (name/title/lore/tips/abilities). También se puede enviar por header Accept-Language. Default: en"
          },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "Número de página (por defecto: 1)" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 }, "description": "Cantidad por página (por defecto: 20)" }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaginatedChampionSummaries" } } }
          },
          "400": { "description": "Error de validación" }
        }
      },
      "post": {
        "tags": ["Campeones"],
        "summary": "Crear campeón custom (protegido)",
        "security": [{ "sessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateChampionBody" } } }
        },
        "responses": {
          "201": { "description": "Creado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Champion" } } } },
          "400": { "description": "Error de validación", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "No autorizado" }
        }
      }
    },
...
    "/api/v1/champions/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "ID del campeón (ej., aatrox, ahri, MyUniqueChamp)" }
      ],
      "get": {
        "tags": ["Campeones"],
        "summary": "Obtener detalles por ID",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "en", "enum": ["en", "es"] },
            "description": "Idioma de textos del campeón. Default: en. Alternativa: header Accept-Language"
          }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Champion" } } } },
          "400": { "description": "Error de validación" },
          "404": { "description": "No encontrado" }
        }
      },
      "patch": {
        "tags": ["Campeones"],
        "summary": "Actualizar campeón custom (protegido + ownership)",
        "security": [{ "sessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateChampionBody" } } }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Error de validación" },
          "401": { "description": "No autorizado" },
          "403": { "description": "Forbidden (seed o no owner)" },
          "404": { "description": "No encontrado" }
        }
      },
      "delete": {
        "tags": ["Campeones"],
        "summary": "Eliminar campeón custom (protegido + ownership)",
        "security": [{ "sessionAuth": [] }],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Error de validación" },
          "401": { "description": "No autorizado" },
          "403": { "description": "Forbidden (seed o no owner)" },
          "404": { "description": "No encontrado" }
        }
      }
    },
...
  }
}